+++ packages/mg-components/src/components/molecules/inputs/mg-input-select/mg-input-select.tsx
@@ -3,7 +3,7 @@
 import { Component, Element, Event, h, Prop, State, EventEmitter, Watch, Method } from '@stencil/core';
 import { MgInput } from '../MgInput';
 import { Handler, Width } from '../MgInput.conf';
-import { ClassList, allItemsAreString, isValidString } from '../../../../utils/components.utils';
+import { ClassList, allItemsAreString, isValidString } from '@mgdis/stencil-helpers';
 import { initLocales } from '../../../../locales';
 import { SelectOption, OptGroup } from './mg-input-select.conf';