+++ packages/stencil-helpers/src/locale/index.spec.ts
@@ -1,9 +1,14 @@
-import { localeCurrency, localeNumber, localeDate, getLocaleMessages, dateToString, getLocaleDatePattern } from './locale.utils';
+import { describe, expect, test } from 'vitest';
+import { localeCurrency, localeNumber, localeDate, defineLocales, dateToString, getLocaleDatePattern } from './';
 import messagesEn from '../locales/en/messages.json';
 import messagesFr from '../locales/fr/messages.json';
 
-describe('locale.utils', () => {
-  describe.each(['en', 'fr'])('locale.utils: %s', locale => {
+const messagesFr = { lang: 'fr' };
+const messagesEn = { lang: 'en' };
+const messages = { en: messagesEn, fr: messagesFr };
+
+describe('locale', () => {
+  describe.each(['en', 'fr'])('locale: %s', locale => {
     describe('localeCurrency', () => {
       test('Should Format number to the locale currency', () => {
         const formatedCurrency = localeCurrency(1234567890.12, locale, 'EUR');
@@ -48,18 +53,28 @@ describe('locale.utils', () => {
       });
     });
 
-    describe('getLocaleMessages', () => {
-      const messages = { en: messagesEn, fr: messagesFr };
-      test('Should return default locale', () => {
-        const locales = getLocaleMessages(document.createElement('div'), messages, 'en');
-        expect(locales.locale).toEqual('en');
+    describe('defineLocales', () => {
+      test('Should return navigator locale', () => {
+        const locales = defineLocales(messages, locale)(document.createElement('div'));
+        expect(locales.locale).toEqual('en-US');
         expect(locales.messages).toMatchObject(messagesEn);
+      });
+
+      test('Should return default locale', () => {
+        const navigatorBackup = window.navigator;
+        globalThis.window.navigator = [];
+        const locales = defineLocales(messages, locale)(document.createElement('div'));
+        expect(locales.locale).toEqual(locale);
+        expect(locales.messages).toMatchObject({ lang: locale });
+
+        // reset default navigator default value
+        globalThis.window.navigator = navigatorBackup;
       });
 
       test('Should return matching locale', () => {
         const div = document.createElement('div');
         div.lang = 'fr-FR';
-        const locales = getLocaleMessages(div, messages, 'en');
+        const locales = defineLocales(messages, locale)(div);
         expect(locales.locale).toEqual(div.lang);
         expect(locales.messages).toMatchObject(messagesFr);
       });
@@ -67,9 +82,9 @@ describe('locale.utils', () => {
       test('Should return default locale messages when requested does not exist', () => {
         const div = document.createElement('div');
         div.lang = 'ca';
-        const locales = getLocaleMessages(div, messages, 'en');
+        const locales = defineLocales(messages, locale)(div);
         expect(locales.locale).toEqual(div.lang);
-        expect(locales.messages).toMatchObject(messagesEn);
+        expect(locales.messages).toMatchObject({ lang: locale });
       });
     });
   });