+++ packages/stencil-helpers/src/storybook/index.conf.ts
@@ -1,0 +1,4 @@
+/**
+ * Define compents default args type
+ */
+export type ArgsType = { [x: string]: unknown };