+++ scripts/import-docs.sh
@@ -7,7 +7,7 @@ dest_dir="./build/docs"
 rm -rf "$dest_dir"
 
 # Import packages markdown
-for file in $(find "$packages_dir" -maxdepth 2 -name "*.md"); do
+for file in $(find "$packages_dir" -maxdepth 3 -name "*.md"); do
   relative_path="${file#$packages_dir}"
   destination="$dest_dir/packages$relative_path"
   mkdir -p "$(dirname "$destination")"